fix: fix unit test script
This commit fixes an issue with the unit test script where it
could not be run without Mocha installed globally since there was
a bug where the alias wasn't working, but if Mocha was installed
globally, it would use that instance to run the tests. However, if
Mocha is not installed globally, then it would fail because it
pointed to the local instance of Mocha using a relative path that
was in direct conflict with the relative path used to run the local
instance of pm2. The npx command fixes these issues by finding the
local instance of a npm binary regardless of what the current
directory is.